+++ App/app/menu.h
@@ -26,6 +26,7 @@
 extern uint8_t gUnlockAllTxConfCnt;
 
 int MENU_GetLimits(uint8_t menu_id, int32_t *pMin, int32_t *pMax);
+uint16_t MENU_BatCalLowPreset(void);
 void MENU_AcceptSetting(void);
 void MENU_ShowCurrentSetting(void);
 void MENU_StartCssScan(void);